Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -2,8 +2,8 @@
 
 
 int main() {
-    int a,b,c;
-    scanf("%d %d %d",&a,&b,&c);
+    double a,b,c;
+    scanf("%lf %lf %lf",&a,&b,&c);
     int avg=a+b+c/3;
     printf("Average: %f",avg);
     return 0;
